+++ activerecord/lib/active_record/connection_adapters/postgresql/schema_statements.rb
@@ -273,9 +273,9 @@ module ActiveRecord
         def default_sequence_name(table_name, pk = nil) #:nodoc:
           result = serial_sequence(table_name, pk || 'id')
           return nil unless result
-          result.split('.').last
+          Utils.extract_schema_qualified_name(result)
         rescue ActiveRecord::StatementInvalid
-          "#{table_name}_#{pk || 'id'}_seq"
+          PostgreSQL::Name.new(nil, "#{table_name}_#{pk || 'id'}_seq")
         end
 
         def serial_sequence(table, column)