+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   PostgreSQL `default_sequence_name` respects schema. Fixes #7516.
+
+    *Yves Senn*
+
 *   Fixed `columns_for_distinct` of postgresql adapter to work correctly
     with orders without sort direction modifiers.